# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "SimbaTracking",
        "description": "SimbaTracking"
    },
    "paths": {
...
        "/settings/get-cms": {
            "get": {
                "x-controller-name": "SettingsController",
                "x-operation-name": "GetCMS",
                "tags": ["SettingsController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetCMS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [],
                "operationId": "SettingsController.GetCMS"
            }
        },
        "/settings/get-settings": {
            "get": {
                "x-controller-name": "SettingsController",
                "x-operation-name": "GetSettings",
                "tags": ["SettingsController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetSettings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [],
                "operationId": "SettingsController.GetSettings"
            }
        },
...
        "/job/cancel-job": {
            "post": {
                "x-controller-name": "JobController",
                "x-operation-name": "CancelJob",
                "tags": ["JobController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CancelJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "JobController.CancelJob"
            }
        },
        "/job/pickedup-job": {
            "post": {
                "x-controller-name": "JobController",
                "x-operation-name": "PickedupJob",
                "tags": ["JobController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "PickedupJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "driver_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "JobController.PickedupJob"
            }
        },
...
    },
...
    "servers": [{
        "url": "http://localhost:1330/api"
    }, {
        "url": "http://15.188.211.169:1330/api"
    }]
}